Add shell for OpenCL CNN implementation
The entire project is separated into an abstract library that define everything needed to construct a generic CNN.
For every specific implementation of the CNN, one is intended to create a new library such as the OpenCLLibrary.
In this case one could create a library using various technology such as CUDA, OpenCL, MPI etc... that can be loaded
dynamically depending on the environment the user is running on.