Add support for multiple input formats to the same CNN
The architecture of the network is to be highly optimized in regards to both memory usage and computations.
The layers of the network is supposed to adapt themselves to the input on creation.
Afterward, they will be immutable with regards to input and output and this is all for performance reasons.
With the current modifications, the network can support multiple input / output dimensions using the same network parameters.