Max pooling pass the gradient test
The horrible bug I was searching was indeed no bug. It was due to the fact that max pooling can have
strongly non-linear behavior which means that we need to use a very small value h when calculating the finite difference.

For max-pooling, if the change in the weights are too big, then we can get completely different max values in the layers which may imply catastrophic behavior.